+++ packages/react/src/dotlottie-worker.tsx
@@ -7,9 +7,12 @@ import type { JSX } from 'react';
 import { BaseDotLottieReact } from './base-dotlottie-react';
 import type { BaseDotLottieProps } from './base-dotlottie-react';
 
-export type DotLottieWorkerReactProps = Omit<BaseDotLottieProps<DotLottieWorker>, 'createDotLottie'>;
+export type DotLottieWorkerReactProps = Omit<BaseDotLottieProps<DotLottieWorker>, 'createDotLottie' | 'workerUrl'> & {
+  workerUrl: string;
+};
 
-const createDotLottieWorker = (config: Config & { workerId?: string }): DotLottieWorker => new DotLottieWorker(config);
+const createDotLottieWorker = (config: Config & { workerId?: string; workerUrl: string }): DotLottieWorker =>
+  new DotLottieWorker(config);
 
 export const DotLottieWorkerReact = (props: DotLottieWorkerReactProps): JSX.Element => {
   return <BaseDotLottieReact {...props} createDotLottie={createDotLottieWorker} />;